requirements.txt
streamlit==1.45.0
pandas==2.2.3
python-dotenv==1.0.1
pymongo==4.12.1